cleanup of start and stop of application
