+++ main.go
@@ -2,18 +2,24 @@ package main
 
 import (
 	"os"
+	"os/signal"
+	"syscall"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
 	"github.com/ilgianlu/tagyou/api"
 	"github.com/ilgianlu/tagyou/conf"
-	mq "github.com/ilgianlu/tagyou/mqtt"
+	"github.com/ilgianlu/tagyou/mqtt"
 	"github.com/ilgianlu/tagyou/persistence"
+	"github.com/ilgianlu/tagyou/routers"
 	dotenv "github.com/joho/godotenv"
 )
 
 func main() {
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, os.Interrupt, syscall.SIGTERM)
+
 	err := loadEnv()
 	if err != nil {
 		log.Fatal().Err(err).Msg("Could not load env")
@@ -29,10 +35,14 @@ func main() {
 
 	p := persistence.SqlPersistence{}
 	p.Init()
-	defer p.Close()
+
+	router := routers.NewSimple()
 
 	go api.StartApi(os.Getenv("API_PORT"))
-	mq.Start(os.Getenv("LISTEN_PORT"), os.Getenv("WS_PORT"))
+	go mqtt.StartWebSocket(os.Getenv("WS_PORT"), router)
+	go mqtt.StartMQTT(os.Getenv("LISTEN_PORT"), router)
+
+	<-c
 }
 
 func loadEnv() error {
