+++ mqtt/tcp_socket.go
@@ -50,14 +50,7 @@ func handleTcpConnection(router routers.Router, conn net.Conn) {
 		err := scanner.Err()
 		if err != nil {
 			if err, ok := err.(net.Error); ok && err.Timeout() {
-				// socket up but silent
-				log.Debug().Msgf("[MQTT] keepalive of %d seconds not respected!", session.KeepAlive*2)
-				if session.GetClientId() != "" {
-					log.Debug().Msgf("[MQTT] (%s:%d) will due to keepalive not respected!", session.GetClientId(), session.LastConnect)
-					willEvent(&session, events)
-					disconnectClient(&session, events)
-					return
-				}
+				onSocketUpButSilent(&session, events)
 			}
 		}
 
@@ -81,6 +74,49 @@ func handleTcpConnection(router routers.Router, conn net.Conn) {
 
 		events <- &p
 	}
+}
+
+func packetSplit(session *model.RunningSession, events chan<- *packet.Packet) func(b []byte, atEOF bool) (int, []byte, error) {
+	return func(b []byte, atEOF bool) (advance int, token []byte, err error) {
+		if len(b) == 0 && atEOF {
+			wasConnected := onSocketDownClosed(session, events)
+			if wasConnected {
+				return 0, nil, nil
+			}
+			return 0, b, bufio.ErrFinalToken
+		}
+		pb, err := packet.ReadFromByteSlice(b)
+		if err != nil {
+			if !atEOF {
+				return 0, nil, nil
+			}
+			log.Debug().Msgf("[MQTT] error reading bytes - session: %s : %s", session.GetClientId(), err.Error())
+			return 0, pb, bufio.ErrFinalToken
+		}
+		return len(pb), pb, nil
+	}
+}
+
+func onSocketUpButSilent(session *model.RunningSession, events chan<- *packet.Packet) bool {
+	log.Debug().Msgf("[MQTT] keepalive of %d seconds not respected!", session.KeepAlive*2)
+	if session.GetClientId() != "" {
+		log.Debug().Msgf("[MQTT] (%s:%d) will due to keepalive not respected!", session.GetClientId(), session.LastConnect)
+		willEvent(session, events)
+		disconnectClient(session, events)
+		return true
+	}
+	return false
+}
+
+func onSocketDownClosed(session *model.RunningSession, events chan<- *packet.Packet) bool {
+	log.Debug().Msgf("[MQTT] socket down closed!")
+	if session.GetClientId() != "" {
+		log.Debug().Msgf("[MQTT] (%s:%d) will due to socket down!", session.GetClientId(), session.LastConnect)
+		willEvent(session, events)
+		disconnectClient(session, events)
+		return true
+	}
+	return false
 }
 
 func willEvent(session *model.RunningSession, e chan<- *packet.Packet) {
